+++ include/nfx/Gl.h
@@ -8,6 +8,7 @@
  */
 
 #include "nfx/graphics/gl/core/buffers/Buffer.h"
+#include "nfx/graphics/gl/core/buffers/VertexArray.h"
 #include "nfx/graphics/gl/core/shaders/ShaderProgram.h"
 #include "nfx/graphics/gl/core/textures/Texture2D.h"
 #include "nfx/graphics/gl/core/Context.h"